// src/gen/types/GetApiActivitiesTrainer.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/


export type GetApiActivitiesTrainerQueryParams = {
    /**
     * @type number | undefined
    */
    limit?: number;
    /**
     * @type number | undefined
    */
    offset?: number;
};

export const payloadDirectionEnum2 = {
    "up": "up",
    "down": "down"
} as const;

export type PayloadDirectionEnum2Key = (typeof payloadDirectionEnum2)[keyof typeof payloadDirectionEnum2];

/**
 * @description Response for status 200
*/
export type GetApiActivitiesTrainer200 = {
    /**
     * @type array
    */
    activities: ({
        /**
         * @type string
        */
        id: string;
        /**
         * @type string
        */
        clientId: string;
        /**
         * @type string
        */
        clientName: string;
        /**
         * @type string
        */
        createdAt: string;
        /**
         * @type string
        */
        type: string;
        /**
         * @type object
        */
        payload: {
            /**
             * @type string | undefined
            */
            workoutName?: string;
            /**
             * @type number
            */
            duration: number;
            /**
             * @type number
            */
            weight: number;
            /**
             * @type number
            */
            series: number;
        };
    } | {
        /**
         * @type string
        */
        id: string;
        /**
         * @type string
        */
        clientId: string;
        /**
         * @type string
        */
        clientName: string;
        /**
         * @type string
        */
        createdAt: string;
        /**
         * @type string
        */
        type: string;
        /**
         * @type object
        */
        payload: {
            /**
             * @type number
            */
            weight: number;
            /**
             * @type number
            */
            weightDifference: number;
            /**
             * @type string
            */
            direction: PayloadDirectionEnum2Key;
        };
    } | {
        /**
         * @type string
        */
        id: string;
        /**
         * @type string
        */
        clientId: string;
        /**
         * @type string
        */
        clientName: string;
        /**
         * @type string
        */
        createdAt: string;
        /**
         * @type string
        */
        type: string;
        /**
         * @type object
        */
        payload: {
            /**
             * @type string | undefined
            */
            routineName?: string;
        };
    })[];
};

export type GetApiActivitiesTrainerQueryResponse = GetApiActivitiesTrainer200;

export type GetApiActivitiesTrainerQuery = {
    Response: GetApiActivitiesTrainer200;
    QueryParams: GetApiActivitiesTrainerQueryParams;
    Errors: any;
};